Added strong naming capability to Release builds.
Gutted the beginnings of the porting of my older plugin framework code. MEF1 was determined to be too costly in performance, and in the spirit of keeping things fresh, I'll be porting over a variant of it that uses MEF2, in particular the newest System.Composition namespace.

Added the beginnings of a proper logging framework for use in a library. Code will be made compilable as soon as practicable. :)