A separate assembly now handles all classic .NET configuration provisioning, and BadEcho.Odin no longer has ugly System.Configuration.ConfigurationManager dependency!
Extensibility configuration now becomes a concern of the application consuming the Odin library, it is no longer a concern of the library, as is proper.

Default implementations of Extensibility configuration interfaces are provided and will work with common configuration binding systems (i.e. Microsoft's IConfiguration binder).

XML configuration entities finalized, schema constructed, and a simple unit test for a valid config has been added. There is some validation logic that might get some unit testing in the future, but it is pretty low priority atm.

This concludes addition of HostAdapter itself in the framework. All that remains are a base plugin adapters class and tests for HostAdapter and some fictitious call-routable plugins.

Corrected a number of analysis warnings.

We're now ready to implement the public facing side of Odin's Extensibility framework and finish it.